Use docdir the standard way so that it can be overridden with --docdir
The default is determined by the 4th argument (PACKAGE_TARBALL) to
AC_INIT, which itself defaults to the 1st argument (PACKAGE_NAME). That
would currently name the docdir `advancemame` rather than the desired
`advance`. Rather than change the 1st argument, set the 4th argument so
that only the docdir is affected. These variables/macros are not really
used elsewhere anyway.