Use /usr/share/man and /usr/share/doc and not /usr/man and /usr/doc (#11)
* Use docdir the standard way so that it can be overridden with --docdir

The default is determined by the 4th argument (PACKAGE_TARBALL) to
AC_INIT, which itself defaults to the 1st argument (PACKAGE_NAME). That
would currently name the docdir `advancemame` rather than the desired
`advance`. Rather than change the 1st argument, set the 4th argument so
that only the docdir is affected. These variables/macros are not really
used elsewhere anyway.

* Install *.html files to a separate htmldir rather than docdir

Autoconf already sets up htmldir for you. It defaults to the same as
docdir but some distributions set it differently.